+++ tests/pt/blocks/test_repeat.py
@@ -1,21 +1,23 @@
 import unittest
-from unittest.mock import patch
+from unittest.mock import patch, Mock
 import torch as pt
 import torch.nn as ptn
-from swak.pt.blocks import Repeat
+from swak.pt.blocks import Repeat, SkipConnection
 from swak.pt.misc import Identity
 
 
 class TestDefaultAttributes(unittest.TestCase):
 
     def setUp(self):
-        self.repeat = Repeat(Identity())
+        self.block = Mock()
+        self.skip = SkipConnection(self.block)
+        self.repeat = Repeat(self.skip)
 
     def test_has_skip(self):
         self.assertTrue(hasattr(self.repeat, 'skip'))
 
     def test_skip(self):
-        self.assertIsInstance(self.repeat.skip, Identity)
+        self.assertIsInstance(self.repeat.skip, SkipConnection)
 
     def test_has_n_layers(self):
         self.assertTrue(hasattr(self.repeat, 'n_layers'))
@@ -24,12 +26,18 @@ class TestDefaultAttributes(unittest.TestCase):
         self.assertIsInstance(self.repeat.n_layers, int)
         self.assertEqual(2, self.repeat.n_layers)
 
-    def test_has_sequence(self):
-        self.assertTrue(hasattr(self.repeat, 'sequence'))
+    def test_has_blocks(self):
+        self.assertTrue(hasattr(self.repeat, 'blocks'))
 
-    def test_sequence(self):
-        self.assertIsInstance(self.repeat.sequence, ptn.Sequential)
-        self.assertEqual(2, len(self.repeat.sequence))
+    def test_blocks(self):
+        self.assertIsInstance(self.repeat.blocks, ptn.Sequential)
+        self.assertEqual(2, len(self.repeat.blocks))
+
+    def test_has_norm(self):
+        self.assertTrue(hasattr(self.repeat, 'norm'))
+
+    def test_norm(self):
+        self.assertIsInstance(self.repeat.norm, Identity)
 
     def test_has_layers(self):
         self.assertTrue(hasattr(self.repeat, 'layers'))
@@ -47,7 +55,7 @@ class TestDefaultAttributes(unittest.TestCase):
     def test_call_reset_parameters(self):
         with patch.object(Identity, 'reset_parameters') as mock:
             self.repeat.reset_parameters()
-            self.assertEqual(2, mock.call_count)
+            self.assertEqual(3, mock.call_count)
 
     def test_has_new(self):
         self.assertTrue(hasattr(self.repeat, 'new'))
@@ -68,40 +76,98 @@ class TestDefaultAttributes(unittest.TestCase):
 class TestAttributes(unittest.TestCase):
 
     def setUp(self):
-        self.repeat = Repeat(
-            Identity(),
-            4
-        )
+        self.block = Mock()
+        self.skip = SkipConnection(self.block)
+        self.repeat = Repeat(self.skip, 4)
 
     def test_n_layers(self):
         self.assertEqual(4, self.repeat.n_layers)
 
-    def test_sequence(self):
-        self.assertEqual(4, len(self.repeat.sequence))
+    def test_blocks(self):
+        self.assertEqual(4, len(self.repeat.blocks))
+
+    def test_norm_first(self):
+        block = Mock()
+        norm = Mock()
+        norm_cls = Mock(return_value=norm)
+        skip = SkipConnection(block, norm_cls=norm_cls, hello='world')
+        repeat = Repeat(skip, 4)
+        norm_cls.assert_called_with(hello='world')
+        self.assertEqual(6, norm_cls.call_count)
+        self.assertIs(repeat.norm, norm)
+
+    def test_norm_after(self):
+        block = Mock()
+        norm = Mock()
+        norm_cls = Mock(return_value=norm)
+        skip = SkipConnection(
+            block,
+            norm_first=False,
+            norm_cls=norm_cls,
+            hello='world'
+        )
+        repeat = Repeat(skip, 4)
+        norm_cls.assert_called_with(hello='world')
+        self.assertEqual(5, norm_cls.call_count)
+        self.assertIsInstance(repeat.norm, Identity)
 
 
 class TestUsage(unittest.TestCase):
 
-    def setUp(self):
-        self.repeat = Repeat(Identity(), 3)
-
-    def test_skip_called(self):
-        with patch.object(Identity, 'forward') as mock:
-            inp = pt.tensor(4)
-            mock.return_value = inp
-            _ = self.repeat(inp)
-            self.assertEqual(3, mock.call_count)
-            arg1 = mock.call_args_list[0][0][0]
-            arg2 = mock.call_args_list[1][0][0]
-            arg3 = mock.call_args_list[2][0][0]
-            self.assertIs(arg1, inp)
-            self.assertIs(arg2, inp)
-            self.assertIs(arg3, inp)
-
-    def test_no_repetitions(self):
-        repeat = Repeat(Identity(), 0)
-        inp = pt.tensor(4)
+    def test_skip_called_norm_first(self):
+        norm_out = pt.ones(4) * 2.0
+        norm = Mock(return_value=norm_out)
+        skip = Mock()
+        skip.new = Mock(return_value=Identity())
+        skip.norm_cls = Mock(return_value=norm)
+        skip.norm_first = True
+        skip.args = ()
+        skip.kwargs = {}
+        repeat = Repeat(skip, 3)
+        inp = pt.ones(4)
         actual = repeat(inp)
+        self.assertIs(actual, norm_out)
+        norm.assert_called_once_with(inp)
+
+    def test_skip_called_norm_after(self):
+        norm_out = pt.ones(4) * 2.0
+        norm = Mock(return_value=norm_out)
+        skip = Mock()
+        skip.new = Mock(return_value=Identity())
+        skip.norm_cls = Mock(return_value=norm)
+        skip.norm_first = False
+        skip.args = ()
+        skip.kwargs = {}
+        repeat = Repeat(skip, 3)
+        inp = pt.ones(4)
+        actual = repeat(inp)
+        self.assertIs(actual, inp)
+        norm.assert_not_called()
+
+    def test_no_repetitions_norm_first(self):
+        block = Mock()
+        expected = pt.ones(4) * 2
+        norm = Mock(return_value=expected)
+        norm_cls = Mock(return_value=norm)
+        skip = SkipConnection(block, norm_cls=norm_cls)
+        repeat = Repeat(skip, 0)
+        inp = pt.ones(4)
+        block.assert_not_called()
+        actual = repeat(inp)
+        norm.assert_called_once_with(inp)
+        self.assertIs(actual, expected)
+
+    def test_no_repetitions_norm_after(self):
+        block = Mock()
+        expected = pt.ones(4) * 2
+        norm = Mock(return_value=expected)
+        norm_cls = Mock(return_value=norm)
+        skip = SkipConnection(block, norm_first=False, norm_cls=norm_cls)
+        repeat = Repeat(skip, 0)
+        inp = pt.ones(4)
+        block.assert_not_called()
+        actual = repeat(inp)
+        norm.assert_not_called()
         self.assertIs(actual, inp)