+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added progress bar for train (and test) loss evaluation.
+- Added progress bar for train (and test) loss evaluation
+- Add options to always step the optimizer after each batch
+- Added norm-first option to (repeated) skip-connection blocks
 
 ## [0.3.1] - 2024-11-27
 - Removed old parent directory creation logic for checkpoint file